requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
passlib[bcrypt]
PyJWT
python-dotenv
pytest
pytest-asyncio
httpx